requirements.txt
azure-mgmt-logic>=10.0.0
azure-identity>=1.12.0
azure-ai-agents>=1.0.0
azure-mgmt-resource>=21.0.0